requirements.txt
numpy
pycollada
vg>=1.7.0,<2.0
